fix(B-464): split session model and schema
